currency string and some bug fixes

+++ src/main/java/openwrestling/model/gameObjects/financial/BankAccount.java
@@ -7,11 +7,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import openwrestling.model.gameObjects.GameObject;
 import openwrestling.model.gameObjects.Promotion;
-import openwrestling.model.segmentEnum.TransactionType;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
 
 @Builder
 @Setter
@@ -23,14 +18,12 @@ public class BankAccount extends GameObject {
     private long bankAccountID;
     private Promotion promotion;
     @Builder.Default
-    private int funds = 1000000;
+    private long funds = 1000000;
 
 
-    public void setFunds(int income) {
+    public void setFunds(long income) {
         funds = income;
     }
-
-
 
 
 }
